Updated route comments and refactored contact, exports, and ping routes (#182)
