+++ controllers/exports.controller.js
@@ -1,0 +1,8 @@
+const { exportURL } = require("../config/index");
+
+const exportListingCalendar = async (req, res) => {
+  const file = `./exports/${req.params.filename}`;
+  res.download(file);
+};
+
+module.exports = { exportListingCalendar };
