+++ routes/accountVerification.js
@@ -5,6 +5,10 @@ const router = express.Router();
 const { requireUserAuth } = require("../utils");
 const AccountVerification = require("../controllers/accountVerification.controller.js");
 
+/*
+  REQ USER: email, _id, name
+  DESCRIPTION: sends verification email to signed in user (indicated by bearer token)
+*/
 router.post(
   "/sendVerificationEmail",
   requireUserAuth,
